# Nest/lec3/lab3/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The order has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new order",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [
          {
            "name": "paymentMethod",
            "required": false,
            "in": "query",
            "description": "Filter by payment method type",
            "schema": {
              "example": "Cash",
              "type": "string",
              "enum": [
                "Cash",
                "Visa"
              ]
            }
          },
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "description": "Filter by client owner ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching orders.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all orders with optional filter options",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The found order record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an order by its ID",
        "tags": [
          "orders"
        ]
      },
      "put": {
        "operationId": "OrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated order record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an existing order",
        "tags": [
          "orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The order record has been successfully deleted."
          },
          "404": {
            "description": "Order not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an order by its ID",
        "tags": [
          "orders"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentialsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully signed up"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/signin": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginVM"
                }
              }
            }
          },
          "401": {
            "description": "Login details invalid"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new product",
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all products",
        "tags": [
          "products"
        ]
      }
    }
  },
  "info": {
    "title": "Orders API",
    "description": "The Orders API description with TypeORM and PostgreSQL",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "auth",
      "description": ""
    },
    {
      "name": "products",
      "description": ""
    },
    {
      "name": "orders",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "longitude": {
            "type": "number",
            "description": "The longitude coordinates",
            "example": 23.21
          },
          "latitude": {
            "type": "number",
            "description": "The latitude coordinates",
            "example": 31.01
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "Cash",
              "Visa"
            ],
            "description": "The payment method type",
            "example": "Visa"
          },
          "productIds": {
            "description": "The list of product IDs included in this order",
            "example": [
              "uuid-1",
              "uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "longitude",
          "latitude",
          "paymentMethod",
          "productIds"
        ]
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the order (UUID)"
          },
          "amount": {
            "type": "number",
            "description": "The order amount"
          },
          "longitude": {
            "type": "number",
            "description": "The longitude of location coordinate"
          },
          "latitude": {
            "type": "number",
            "description": "The latitude of location coordinate"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "Cash",
              "Visa"
            ],
            "description": "The payment method"
          }
        },
        "required": [
          "id",
          "amount",
          "longitude",
          "latitude",
          "paymentMethod"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "The order amount"
          },
          "longitude": {
            "type": "number",
            "description": "The longitude coordinates"
          },
          "latitude": {
            "type": "number",
            "description": "The latitude coordinates"
          },
          "clientId": {
            "type": "string",
            "description": "The client owner ID"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "Cash",
              "Visa"
            ],
            "description": "The payment method type"
          },
          "productIds": {
            "description": "The list of product IDs included in this order",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AuthCredentialsDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The unique username",
            "example": "mazen123"
          },
          "email": {
            "type": "string",
            "description": "The unique email address",
            "example": "mazen@example.com"
          },
          "password": {
            "type": "string",
            "description": "The secure password",
            "example": "Mazen_Pass123"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username",
            "example": "mazen123"
          },
          "password": {
            "type": "string",
            "description": "The password",
            "example": "Mazen_Pass123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginVM": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT Bearer Access Token"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The product name",
            "example": "Samsung S24 Ultra"
          },
          "price": {
            "type": "number",
            "description": "The product price",
            "example": 1200
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the product (UUID)"
          },
          "name": {
            "type": "string",
            "description": "The name of the product"
          },
          "price": {
            "type": "number",
            "description": "The price of the product"
          }
        },
        "required": [
          "id",
          "name",
          "price"
        ]
      }
    }
  }
}